+++ packages/backend/src/index.js
@@ -1,8 +1,9 @@
-import { startServer } from "app";
+import { startServer } from "App";
 import { startWorkers } from "workers";
 import env from "environment";
 
 startServer();
+
 if (env("enable_workers")) {
   startWorkers();
 } else {
